requirements.txt
# requirements.txt
python-telegram-bot==22.2
python-dotenv==1.1.1
pydub==0.25.1
openai-whisper==20240930
llama-cpp-python==0.3.11
transformers==4.53.1
accelerate==1.8.1
torch==2.7.1
huggingface-hub==0.33.2